增量打jar包，打dex在customJavaC中完成

+++ fastdex-gradle/src/main/groovy/fastdex/build/transform/FastdexDexBuilderTransform.groovy
@@ -20,6 +20,9 @@ class FastdexDexBuilderTransform extends TransformProxy {
     @Override
     void transform(TransformInvocation transformInvocation) throws TransformException, IOException, InterruptedException {
         if (fastdexVariant.hasDexCache) {
+            if(fastdexVariant.configuration.useCustomCompile){
+                return
+            }
             project.logger.error("\n==fastdex patch transform start,we will generate dex file")
             if (fastdexVariant.projectSnapshoot.diffResultSet.isJavaFileChanged()) {
                 FileUtils.deleteDir(streamOutputFolder)
